backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
supabase==2.4.2
pydantic[email]>=2.9.0
slowapi==0.1.9
python-dotenv==1.0.1
